text-maxi `listTab` -> `ListOptionsControl`
