+++ src/blocks/text-maxi/components/index.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
+export { default as ListOptionsControl } from './list-options-control';
